Iniciando frontend

+++ .docker/entrypoint.sh
@@ -1,10 +1,20 @@
 #!/bin/bash
 
-chmod -R 777 /var/www/storage && chmod -R 777 /var/www/bootstrap/cache
+chmod -R 777 /var/www/backend/storage && chmod -R 777 /var/www/backend/bootstrap/cache
 
 #On error no such file entrypoint.sh, execute in terminal - dos2unix .docker\entrypoint.sh
-cp .env.testing.exemple .env.testing
-cp .env.example .env
+### FRONT-END
+npm config set cache /var/www/.npm-cache --global
+cd /var/www/frontend && npm install && cd ..
+
+### BACK-END
+cd backend 
+if [ ! -f ".env.testing"]; then
+    cp .env.testing.exemple .env.testing
+fi
+if [ ! -f ".env"]; then
+    cp .env.example .env
+fi
 composer install
 php artisan key:generate
 php artisan migrate
